Updated 2D renderer
Updated an issue for sprite importing v1 files.  The sprite would load
the texture prematurely in the process and this caused the deferred
sprite binding to miss that the sprite was a v1 sprite.  This caused the
system to mess up the texture coordinates (v1 uses pixel coordinates, v2
uses texel coordinates) because it does not know that it has to convert
them to texel space.