Refactoring
Updated the application settings API with a more robust comment for the
class, and refactored its internals to use .NET 4.0/4.5 specific
functionality. Also added the ability to use arrays and dictionaries
with the settings class.

Corrected a bug in the property reflection code where if <T> was a
parent type of the declared type on the property, the expression would
fail to run/compile.