Refactored chunk file objects
Changed the chunk file objects to be more representative of the original
design that was envisioned for this format. The old chunk file format
objects have been retired.

This means the chunk file format has changed significantly, and files
written with the old chunk file format will no longer work.