Added new error code for GorgonException
Change behaviour of Begin() to ensure that it is only called once. It will now throw an exception more than 1 call is made without pairing it with an End() call.  This is done to ensure that apps don't forget to call their End() method and the end of a frame.