Updated core graphics API to use better means of detecting state changes between draw calls. This has improved performance a little, and has cut down on some unnecessarily complicated code.
Fixed a bug in the sampler states collection where assigning NULL to a sampler state caused an exception.

Restored draw call cache counts for pool allocator.

Fixed shader resource view list to return the dirty slots from 0..N for the native backing array.  Previously it was returning slots M..N where M was the index of the slot in the GorgonArray, which was incorrect.