Image editor update (WIP).
Updated some nuget references.
Reorganized files.
Added undo capability (redo is WIP)
Added pixel format conversion.
Added more supported formats to BMP and JPeg codecs.
Refactored view model to not require so many parameters on initialization (to keep with SRP as much as possible).
Added image I/O service for handling input/output of image data on the disk.
Fixed bug in image code where converting from sRGB to non-sRGB format would fail when it shouldn't have.
Added functionality to prompt when creating a project, opening a project or closing the application and content is open with unsaved changes.
Updated image editor to set state as modified when converting pixel formats.
Added functionality to allow writing into content files.
Added functionality to save images in the image editor.
Added functionality to allow block compression (BC1-BC7) for images.
Added functionality to refresh preview thumbnails.
Added method to swap chain to allow retrieval of backbuffer as an IGorgonImage or GorgonTexture2D.
Updated GorgonSelectablePanel to fire an event when it was being rendered to a GDI+ bitmap.
Added method to ContentBaseControl to render the contents of a swap chain to a GDI+ bitmap when DrawToBitmap is called.
Fixed annoying bug where graphics rendered to swap chain would not show up when the transparent overlay was active.